backend/requirements.txt
Flask
pymongo[srv]
flask-cors
pandas
scipy
python-dotenv